user registration

+++ base/views.py
@@ -1,6 +1,9 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
 from .models import Room, Topic
@@ -12,8 +15,11 @@ from .form import RoomForm
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if(request.user.is_authenticated):
+        return redirect('home')
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -26,7 +32,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password does not exist')
         
-    context = {}
+    context = {'page': page}
     
     return render(request, 'base/login_register.html', context)
 
@@ -35,6 +41,24 @@ def logoutUser(request):
     logout(request)
     return redirect('home')
 
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+    if(request.method == 'POST'):
+        form = UserCreationForm(request.POST)
+        if(form.is_valid()):
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An occurred during registration')
+    
+    context = {'form': form, 'page': page}
+    
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -50,6 +74,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -60,9 +85,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if(request.user != room.host):
+        return HttpResponse('You are allowed here!!')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -71,8 +100,12 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if(request.user != room.host):
+        return HttpResponse('You are allowed here!!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
